+++ Source/EventFlow/ValueObjects/ValueObject.cs
@@ -21,7 +21,6 @@
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 using System;
-using System.Collections;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
@@ -37,31 +36,16 @@ namespace EventFlow.ValueObjects
         {
             if (ReferenceEquals(this, obj)) return true;
             if (ReferenceEquals(null, obj)) return false;
-            if (obj.GetType() != GetType()) return false;
-            return obj.GetHashCode() == GetHashCode();
+            if (GetType() != obj.GetType()) return false;
+            var other = obj as ValueObject;
+            return other != null && GetEqualityComponents().SequenceEqual(other.GetEqualityComponents());
         }
 
         public override int GetHashCode()
         {
             unchecked
             {
-                return GetEqualityComponents().Aggregate(17, (current, obj) => current*23 + CalculateHashcode(obj));
-            }
-        }
-
-        private static int CalculateHashcode(object obj)
-        {
-            unchecked
-            {
-                if (obj == null)
-                {
-                    return 0;
-                }
-
-                var enumerable = obj as IEnumerable;
-                return enumerable == null
-                    ? obj.GetHashCode()
-                    : enumerable.Cast<object>().Aggregate(17, (current, o) => current*23 + CalculateHashcode(o));
+                return GetEqualityComponents().Aggregate(17, (current, obj) => current * 23 + (obj != null ? obj.GetHashCode() : 0));
             }
         }