+++ RELEASE_NOTES.md
@@ -1,7 +1,7 @@
 ### New in 0.4
 
 * Feature: Release notes added to NuGet packages
-* Fixed: Issue regarding `ValueObject` containing lists
+* Fixed: Unchecked missing in `ValueObject` when claculating hash
 
 ### New in 0.3 (released 2015-04-30)